feat: Add artist search functionality to GUI
This commit adds the ability to search for artists in the GUI. It includes a new text entry field for entering the artist's name and a button to initiate the search. The search results are displayed in a combobox, allowing the user to select from the list of found artists. This feature enhances the user experience by providing a convenient way to find and select artists for playlist generation.